+++ lib/pages/debug_page.dart
@@ -1,9 +1,4 @@
-import 'package:arquivolta/actions.dart';
-import 'package:arquivolta/app.dart';
-import 'package:arquivolta/interfaces.dart';
 import 'package:arquivolta/logging.dart';
-import 'package:arquivolta/widgets/page_scaffold.dart';
-import 'package:arquivolta/services/job.dart';
 import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
