+++ lib/pages/install/install_page.dart
@@ -2,7 +2,8 @@ import 'package:arquivolta/actions.dart';
 import 'package:arquivolta/app.dart';
 import 'package:arquivolta/interfaces.dart';
 import 'package:arquivolta/logging.dart';
-import 'package:arquivolta/pages/install_progress_page.dart';
+import 'package:arquivolta/pages/install/install_progress.dart';
+import 'package:arquivolta/pages/install/install_prompt.dart';
 import 'package:arquivolta/widgets/paged_view.dart';
 import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
@@ -79,80 +80,6 @@ class InstallPage extends HookWidget implements Loggable {
           ),
         ],
       ),
-    );
-  }
-}
-
-class InstallPrompt extends HookWidget implements Loggable {
-  final void Function(String distro, String user, String password)
-      onPressedInstall;
-
-  const InstallPrompt({Key? key, required this.onPressedInstall})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    final distro = useTextEditingController(text: 'arch-foobar');
-    final user = useTextEditingController();
-    //final password = useTextEditingController();
-    //final passwordHidden = useState(true);
-
-    // Reevaluate shouldEnableButton whenever any of the text changes
-    [
-      distro,
-      user, /*password*/
-    ].forEach(useValueListenable);
-
-    final shouldEnableButton = distro.text.length > 1 &&
-        user.text.length > 1; //&& password.text.length > 1;
-
-    final onPress = useCallback(
-      () => onPressedInstall(
-        distro.text,
-        user.text,
-        '', //password.text,
-      ),
-      [
-        distro,
-        user, /*password*/
-      ],
-    );
-
-    return Flex(
-      direction: Axis.vertical,
-      children: [
-        TextBox(
-          controller: distro,
-          header: 'WSL Distro Name',
-        ),
-        TextBox(
-          controller: user,
-          header: 'Linux Username',
-        ),
-        /*
-        TextBox(
-          controller: password,
-          header: 'Password',
-          obscureText: passwordHidden.value,
-          suffix: IconButton(
-            icon: Icon(
-              passwordHidden.value ? FluentIcons.lock : FluentIcons.unlock,
-            ),
-            onPressed: () => passwordHidden.value = !passwordHidden.value,
-          ),
-        ),
-        */
-        Align(
-          alignment: Alignment.centerRight,
-          child: Padding(
-            padding: const EdgeInsets.symmetric(vertical: 8),
-            child: FilledButton(
-              onPressed: shouldEnableButton ? onPress : null,
-              child: const Text('Install'),
-            ),
-          ),
-        ),
-      ],
     );
   }
 }
